Bugfix #9673: Origin header is always sent from WebSocket client
